Make browserify options reusable
