+++ middleware/browserify.js
@@ -1,26 +1,8 @@
 const browserify = require('browserify-middleware')
-
-const nodeModulesToAvoidBabelifying = [
-  'lodash',
-  'lunr',
-  'prettydate'
-]
-
-const excludeRegex = new RegExp(`/node_modules/(${nodeModulesToAvoidBabelifying.join('|')})`)
+const browserifyOptions = require('./browserify-opts')
 
 function babelifyMiddleware (entry) {
-  return browserify(entry, {
-    transform: [
-      ['babelify', {
-        global: true,
-        exclude: excludeRegex,
-        presets: [
-          ['@babel/preset-env', { targets: '> 0.25%, not dead' }]
-        ]
-      }],
-      'brfs'
-    ]
-  })
+  return browserifyOptions(browserify)(entry)
 }
 
 module.exports = babelifyMiddleware
